Restructured and renamed Account Notes controller and services
