+++ src/main/java/com/salessparrow/api/services/accountNotes/GetNotesListService.java
@@ -1,4 +1,4 @@
-package com.salessparrow.api.services.accounts;
+package com.salessparrow.api.services.accountNotes;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
